refactor: wip enhanced storage api
