+++ src/iframe/storage/unpartitioned-storage/local-storage.ts
@@ -1,0 +1,7 @@
+import { UnpartitionedStorage } from "./unpartitioned-storage";
+
+export class LocalStorage extends UnpartitionedStorage {
+  constructor() {
+    super({ area: "local" });
+  }
+}
