+++ src/utils/embedded/embedded.utils.ts
@@ -1,7 +1,8 @@
 import { createSupabaseClient, createTRPCClient } from "embed-api";
 import { jwtDecode } from "jwt-decode";
 import { IS_EMBEDDED_APP } from "~utils/embedded/embedded.constants";
-import { LocalStorage } from "~iframe/local-storage/local-storage";
+import { LocalStorage } from "~iframe/storage/unpartitioned-storage/local-storage";
+import { searchParams, ancestorOrigin, isInsideIframe } from "./iframe.utils";
 
 // Then, its tRPC client will be initialized with the following headers:
 // - authorization (getAuthTokenHeader / setAuthTokenHeader)
@@ -38,13 +39,6 @@ const PARAM_CLIENT_ID = "client-id";
 const PARAM_SERVER_BASE_URL = "server-base-url";
 const PARAM_ANCESTOR_ORIGIN = "ancestor-origin";
 
-const { search = "", ancestorOrigins = [] } = IS_EMBEDDED_APP
-  ? document.location
-  : {};
-
-const searchParams = new URLSearchParams(search);
-const ancestorOrigin = ancestorOrigins[ancestorOrigins.length - 1];
-
 const EMBEDDED_CLIENT_ID =
   searchParams.get(PARAM_CLIENT_ID) ||
   EMBEDDED_ENV_VARS.DEFAULT_EMBEDDED_CLIENT_ID;
@@ -79,16 +73,6 @@ async function handleAuthError() {
     window.location.reload();
   } catch (err) {
     console.error("Error signing out:", err);
-  }
-}
-
-export function isInsideIframe(): boolean {
-  try {
-    return window.self !== window.top || !!ancestorOrigin;
-  } catch (e) {
-    // If we can't access window.top due to cross-origin restrictions,
-    // we're definitely in an iframe
-    return true;
   }
 }
