+++ src/utils/embedded/utils/messages/embedded-messages.utils.ts
@@ -1,8 +1,6 @@
 import { nanoid } from "nanoid";
-import {
-  getEmbeddedAncestorOrigin,
-  isInsideIframe
-} from "~utils/embedded/embedded.utils";
+import { getEmbeddedAncestorOrigin } from "~utils/embedded/embedded.utils";
+import { isInsideIframe } from "~utils/embedded/iframe.utils";
 import type {
   EmbeddedCall,
   EmbeddedMessageId,
